documentation + remove unused function and move detect_syscalls_in_sym_table from syscalls.py to elf_analyser.py
