+++ static/elf_analyser.py
@@ -22,3 +22,13 @@ def get_syscalls_from_symbols(binary, syscalls_set):
     for sect_it in [binary.dynamic_symbols, binary.static_symbols,
                     binary.symbols]:
         detect_syscalls_in_sym_table(sect_it, syscalls_set)
+
+def detect_syscalls_in_sym_table(sect_it, syscalls_set):
+    for s in sect_it:
+        name = s.name
+        name_value = alias_syscalls_map.get(name)
+        if name_value is not None:
+            name = alias_syscalls_map[name]
+
+        if name in syscalls_map:
+            syscalls_set.add(name)
